3D model for the end-cap clip
Basically the signal connection is made directly to the outer rim at
the end of the vidicon tube (I think). My plan for this is to print
this part, which has an opening slightly larger than the cap, and then
to glue in solder wick around the inner part of the opening on this
part. This should make a good connection, I hope.